backend/requirements.txt
# ─── Web Framework ────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0

# ─── Database ─────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1

# ─── Validation & Settings ────────────────────────────────
pydantic==2.8.0
pydantic-settings==2.3.0

# ─── Auth & Security ──────────────────────────────────────
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
python-multipart==0.0.9
cryptography==42.0.7
pyotp==2.9.0
qrcode[pil]==7.4.2
pillow==10.3.0

# ─── Cache & Rate Limiting ────────────────────────────────
redis[asyncio]==5.0.4
slowapi==0.1.9

# ─── LLM Providers ────────────────────────────────────────
anthropic==0.28.0
httpx==0.27.0

# ─── Vector Store ─────────────────────────────────────────
chromadb==0.5.3

# ─── Text Splitting ───────────────────────────────────────
langchain-text-splitters==0.2.2

# ─── PDF ──────────────────────────────────────────────────
pypdf==4.2.0
PyMuPDF==1.24.5

# ─── Word / PowerPoint / Excel ────────────────────────────
python-docx==1.1.2
python-pptx==0.6.23
pandas==2.2.2
openpyxl==3.1.4
tabulate==0.9.0
xlrd==2.0.1

# ─── File handling ────────────────────────────────────────
aiofiles==23.2.1

# ─── Email ────────────────────────────────────────────────
resend==2.6.0
apscheduler==3.10.4


# ─── Logging ──────────────────────────────────────────────
structlog==24.1.0

# ─── Agent Orchestration ──────────────────────────────────
celery[redis]==5.3.6